requirement.txt
aioredis==1.2.0
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
bcrypt==3.1.6
cffi==1.12.2
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
daphne==2.3.0
Django==2.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1
msgpack==0.6.1
pycparser==2.19
pydotplus==2.0.2
PyHamcrest==1.9.0
pyparsing==2.4.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
zope.interface==4.6.0
